Clean up display of length to handle the change made to database

+++ app/src/main/java/com/funjim/fishstory/model/Event.kt
@@ -100,7 +100,7 @@ data class EventSummary(
     val tackleBoxCount: Int,
     val bigFishName: String? = null,
     val bigFishSpecies: String? = null,
-    val bigFishLength: Double? = null,
+    val bigFishLength: Long? = null,
     val mostCaughtName: String? = null,
     val mostCaught: Int? = null
 )
